+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features
 
+1. [#402](https://github.com/influxdata/influxdb-client-js/pull/402): Limit write batch maximum size in bytes.
 1. [#403](https://github.com/influxdata/influxdb-client-js/pull/403): Regenerate APIs from swagger.
 
 ### Bug Fixes